Program3b Assignment Commit
- Merged ProcessProfile into ProcessProfiles
- Added completion order tracking
- Added ProcessProfiles ToString method
