+++ ProcessSchedulingSim/ProcessSchedulingSim.csproj
@@ -44,7 +44,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ProcessProfile\ProcessProfile.cs" />
     <Compile Include="ProcessProfile\ProcessProfiles.cs" />
     <Compile Include="Processor.cs" />
     <Compile Include="Program.cs" />
